refactor code

+++ DTG01/Core/Inc/adc.h
@@ -37,12 +37,12 @@ typedef struct
 	uint32_t vdda_mv;			// VDDA in mv
 	uint32_t vref_int;
 #ifdef DTG02
-	uint32_t in_4_20ma_in[4];		// 4.5mA =>> convert to 45
+	float in_4_20ma_in[4];		// 4.5mA =>> convert to 45
 #else
-	uint32_t in_4_20ma_in[1];		// 4.5mA =>> convert to 45
+	float in_4_20ma_in[1];		// 4.5mA =>> convert to 45
 #endif
 	int32_t temp;
-    int32_t temp_is_valid;
+    uint32_t temp_is_valid;
 } adc_input_value_t;
 
 /**
